\name{HMMEMundirStab-package}
\alias{HMMEMundirStab-package}
\alias{HMMEMundirStab}
\docType{package}
\title{
\packageTitle{HMMEMundirStab}
}
\description{
\packageDescription{HMMEMundirStab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HMMEMundirStab}
\packageIndices{HMMEMundirStab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HMMEMundirStab}

Maintainer: \packageMaintainer{HMMEMundirStab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
